+++ packages/create-amplify/src/amplify_project_creator.ts
@@ -34,6 +34,7 @@ export class AmplifyProjectCreator {
     private readonly initialProjectFileGenerator: InitialProjectFileGenerator,
     private readonly npmInitializedEnsurer: NpmProjectInitializer,
     private readonly tsConfigInitializer: TsConfigInitializer,
+    private readonly projectRoot: string,
     private readonly logger: typeof console = console
   ) {}
 
@@ -68,8 +69,15 @@ export class AmplifyProjectCreator {
     this.logger.log('Scaffolding initial project files...');
     await this.initialProjectFileGenerator.generateInitialProjectFiles();
 
+    const cdCommand =
+      process.cwd() === this.projectRoot
+        ? '`'
+        : `\`cd .${this.projectRoot.replace(process.cwd(), '')}; `;
+
     this.logger.log(
-      `All done! Run \`amplify help\` for a list of available commands. Get started by running \`amplify sandbox\``
+      `All done! 
+Run \`npx amplify help\` for a list of available commands. 
+Get started by running ${cdCommand}npx amplify sandbox\`.`
     );
   };
 }