+++ packages/model-generator/src/index.ts
@@ -1,4 +1,15 @@
 export * from './model_generator.js';
 export * from './create_graphql_document_generator.js';
-export * from './generate_api_code.js';
+export {
+  GenerateApiCodeFormat,
+  GenerateApiCodeModelTarget,
+  GenerateApiCodeStatementTarget,
+  GenerateApiCodeTypeTarget,
+  GenerateModelsOptions,
+  GenerateGraphqlCodegenOptions,
+  GenerateIntrospectionOptions,
+  GenerateOptions,
+  GenerateApiCodeProps,
+  generateApiCode,
+} from './generate_api_code.js';
 export * from './generate_model_introspection_from_schema_uri.js';